requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asarPy==1.0.1
attrs==25.4.0
audioop-lts==0.2.2
discord.py==2.6.4
dotenv==0.9.9
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
mysql-connector-python==9.5.0
propcache==0.4.1
python-dotenv==1.2.1
yarl==1.22.0
